fix: subdividing constants
