+++ src/config/constants/encode.ts
@@ -1,0 +1,3 @@
+export const BASE64 = 'base64';
+export const UTF8 = 'utf8';
+export const HEX = 'hex';
