+++ src/config/constants/filePath.ts
@@ -1,0 +1,4 @@
+export const FILE_PATH = {
+  PRIVATE_KEY_FILE_PATH: 'key/private.key',
+  ADMIN_PRIVATE_KEY_FILE_PATH: 'key/admin-private.key',
+};
