+++ src/config/constants/http.ts
@@ -1,0 +1,14 @@
+export const HTTP_URL = {
+  P2P_GET_QUAKE_HISTORY_URL: 'https://api.p2pquake.net/v2/history',
+  LINE_API_BASE_URL: 'https://api.line.me/',
+  LINE_API_OAUTH_TOKEN_URL: 'https://api.line.me/oauth2/v2.1/token',
+  LINE_API_OAUTH_VERIFY_URL: 'https://api.line.me/oauth2/v2.1/verify',
+  LINE_API_GET_USER_PROFILE_URL: 'https://api.line.me/v2/bot/profile/',
+  LINE_API_PUSH_MESSAGE_URL: 'https://api.line.me/v2/bot/message/push',
+  YAHOO_WEATHER_JP_EARTHQUAKE:
+    'https://typhoon.yahoo.co.jp/weather/jp/earthquake/',
+};
+
+export const HTTP_HEADER = {
+  LINE_SIGNATURE_HEADER: 'x-line-signature',
+};
