+++ src/config/constants/jwt.ts
@@ -1,0 +1,2 @@
+export const JWT = 'JWT';
+export const JWT_FORMAT_COMPACT = 'compact';
