+++ src/config/constants/tableName.ts
@@ -1,0 +1,4 @@
+export const TABLE_NAME = {
+  QUAKE_HISTORY_TABLE_NAME: 'QuakeHistory',
+  CHANNEL_ACCESS_TOKEN_TABLE_NAME: 'ChannelAccessToken',
+};
