+++ src/config/constants/timeZone.ts
@@ -1,0 +1,1 @@
+export const JAPAN_TIMEZONE = 'Asia/Tokyo';
