+++ src/infrastructure/api/line/userApi.ts
@@ -2,16 +2,10 @@ import { UserProfileResponse } from '@line/bot-sdk/dist/messaging-api/model/user
 import { HttpService } from '@nestjs/axios';
 import { Injectable, Logger } from '@nestjs/common';
 import { firstValueFrom } from 'rxjs';
-import { LINE_API_GET_USER_PROFILE_URL } from 'src/config/constants';
+import { HTTP_URL } from 'src/config/constants/http';
+import { LOG_MESSAGES } from 'src/config/logMessages';
 import { IUserApi } from 'src/domain/interfaces/api/line/userApi';
 import { createAuthHeaders } from 'src/domain/useCase/http';
-
-// Log message constants
-const LOG_MESSAGES = {
-  REQUEST_USER_PROFILE: 'Requesting user profile from the LINE Messaging API',
-  FETCH_USER_PROFILE_FAILED:
-    'Failed to fetch user profile from the LINE Messaging API',
-};
 
 /**
  * LINE user API
@@ -34,7 +28,7 @@ export class UserApi implements IUserApi {
   ): Promise<UserProfileResponse> {
     this.logger.log(LOG_MESSAGES.REQUEST_USER_PROFILE);
 
-    const url = `${LINE_API_GET_USER_PROFILE_URL}${userId}`;
+    const url = `${HTTP_URL.LINE_API_GET_USER_PROFILE_URL}${userId}`;
     const headers = createAuthHeaders(channelAccessToken);
 
     try {
